# Single Logger implementation for FairLogger & FairMQLogger