--- a/fairmq/FairMQLogger.h
+++ b/fairmq/FairMQLogger.h
@@ -17,8 +17,4 @@
 
 #include <Logger.h>
 
-using timestamp_t = unsigned long long;
-
-timestamp_t get_timestamp();
-
 #endif /* FAIRMQLOGGER_H_ */