This introduces `Logger` class that is used by:
- `FairMQLogger` (uses `Logger` directly, header just for backwards-compatibility)
- `FairLogger` (minimal wrapper class to maintain some of the special `FairLogger` features and backwards-compatibility).

TODOs:

* [x] Provide all severity and verbosity levels used by both old loggers (with some output enchantments).
* [x] Change all severity names to lowercase. Uppercase is still available for backwards-compatibility.
* [x] Remove usage of `MQLOG` macro, previously used to avoid conflict of both old loggers.
* [x] Make `info` default severity (default in `FairMQLogger` was `DEBUG`) and `low` the default verbosity (default in `FairMQLogger` was `medium`).
* [x] Provide a hook after `FATAL` message, FairLogger uses it to throw an exception and call `gSystem->StackTrace()`.
* [x] Refactor LOG calls that relied on `FairLogger` not adding `endl` automatically.
* [x] Remove `FairLogger::endl` usage from code, which now does basically nothing.
* [x] Removed `TMCThreadLocal` from FairLogger instance. `Logger` is already thread safe.
* [x] Fix FairMQ tests: fix captured stream.
* [x] Fix failing `FairLogger` tests.
* [x] Do more tests with root macros to make sure nothing broke.
* [x] Mark the printf-style interface of FairLogger as deprecated.

The new Logger is backwards-compatible, at least from what I could test so far.