@rbx 2018-01-12
While testing the new implementation I noticed some performance degradation with Boost compared to FairLogger. The runtime of the entire FairRoot test suite almost doubled with the new implementation.

First thing I tried to improve it is to disable all customizations on top of boost to make sure it is nothing that I added on top - the performance remained unchanged.

The discussion [here](https://sourceforge.net/p/boost-log/discussion/710021/thread/47dd83b3/) suggests that the timestamps that boost calculates are not very efficient and also logging of entries that should be hidden (e.g. TRACE&DEBUG calls during INFO severity) takes considerable amount of time (all attributes are calculated even if a log line will not be sinked).

I setup a small test that logs all severities & verbosities followed by 32 million log lines in a "hidden" severity (so they should be skipped as fast as possible).

With default boost log setup test runs in **12 seconds**.

Making a custom calculation for the timestamps in the formatters using `std::chrono` reduced this time to **8 seconds**.

Further I tried to make a custom check in the macro that checks severity before even giving a line to boost - this reduced the time to **1.6 seconds**.

At this point already considerable amount of customizations have been added where not much is left for boost to do (at least from the features that we use), but the runtime of the FairRoot test suite is still almost double.

I tried a custom logger implementation without boost ([branch](https://github.com/rbx/FairRoot/tree/logger-experimental)) just with `cout` and `fstream`. With it the small test completes in just **0.1 seconds**. And the test suite runs as fast as before. It now has all of the features we require.

Here are some additional test results:

|                      | old (two loggers) | new boost       | new custom |
|:---------------------|:-----------------:|:---------------:|:----------:|
| compile time         | 7.1 min           | 9.2 min         | 6.9 min    |
| test suite runtime   | 1.7 min           | **2.9** min     | 1.7 min    |
| small test runtime | -                 | **12 (8, 1.6)** sec | **0.1** sec    |
| bin dir size         | 292.4 MB          | 293.8 MB        | 279.9 MB   |
| lib dir size         | 146.3 MB          | 173.2 MB        | 132.9 MB   |

My guess as to why the runtime of the test suite is so much slower with boost is that cling has to parse all the boost log headers on each execution. But I have not verified this.

Considering these test results (and also the fact that at least one more customization on top of boost would be necessary to provide the required special handling of the FATAL severity) I think we should ditch Boost.Log and use a custom implementation. It has all the features that we use, unless we foresee need of some of the more unique boost log features (log file rotation, ...?).