@rbx 2018-01-16
As discussed I merged my experimental logger branch here.
From what I can tell this is now feature complete.
I updated the TODO list above to reflect all the changes.