@ktf 2018-01-17
I am trying this in ALICE O2 but this is not backward compatible, apparently. I get:

```
sw/BUILD/O2-latest/log
Scanning dependencies of target SubframeBuilderDevice
[ 32%] Building CXX object Utilities/DataFlow/CMakeFiles/SubframeBuilderDevice.dir/src/runSubframeBuilderDevice.cxx.o
/mnt/mesos/sandbox/sandbox/sw/SOURCES/O2/0_O2_DEV_FAIRROOT/0/Framework/Core/src/runDataProcessing.cxx: In function 'int doChild(int, char**, const o2::framework::DeviceSpec&)':
/mnt/mesos/sandbox/sandbox/sw/SOURCES/O2/0_O2_DEV_FAIRROOT/0/Framework/Core/src/runDataProcessing.cxx:250:13: error: 'fair::mq::logger' has not been declared
   fair::mq::logger::ReinitLogger(false);
             ^~~~~~
make[2]: *** [Framework/Core/CMakeFiles/Framework.dir/src/runDataProcessing.cxx.o] Error 1
```

It would be good if changes like this are tested against O2 and announced in the Framework meeting.

Any idea on how to solve the above issue?