@rbx 2018-01-17
@ktf sorry, you're right, I didn't test with O2. The backwards-compatibility is for the LOG macro and FairLogger interfaces - I wasn't aware anyone is using the FairMQ logger methods except in our internal code. Will check more carefully next time.

Anyway the replacement for the above call is `fair::Logger::SetConsoleColor(false);`.