@ktf 2018-01-17
Great, thanks. I am in the process of updating my PR.